+++ Backend/project/src/main/java/com/pac/gestoreeventi/eventsManagement/EventController.java
@@ -20,7 +20,7 @@ public class EventController {
 	}
 
 	@GetMapping(path = "/events/{idEvent}")
-	public Event getEvent(@PathVariable("idEvent") long idEvent){return eventService.getEvent(idEvent);}
+	public Event getEvent(@PathVariable("idEvent") int idEvent){return eventService.getEvent(idEvent);}
 
 	@PostMapping(path = "/events/new")
 	public String insertEvent(@RequestBody EventDTO eventDTO){
@@ -29,5 +29,5 @@ public class EventController {
 	}
 	
 	@DeleteMapping(path = "/events/{idEvent}")
-	public String deleteEvent(@PathVariable("idEvent") Long idEvent){eventService.deleteEvent(idEvent); return "Evento eliminato";}
+	public String deleteEvent(@PathVariable("idEvent") Integer idEvent){eventService.deleteEvent(idEvent); return "Evento eliminato";}
 }
