+++ Backend/project/src/main/java/com/pac/gestoreeventi/eventsManagement/EventRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface EventRepository extends JpaRepository<Event, Long> {
-    Optional<Event> findById(Long id);
+public interface EventRepository extends JpaRepository<Event, Integer> {
+    Optional<Event> findById(Integer id);
 }
